Divided the large main class into several mixins.
