+++ src/logging.py
@@ -1,0 +1,5 @@
+""""""
+
+class Logging:
+    def log(self, msg: str):
+        ...
